calender_api/requirements.txt
Package                  Version
------------------------ --------
asgiref                  3.7.1
cachetools               5.3.0
certifi                  2023.5.7
charset-normalizer       3.1.0
Django                   4.2.1
djangorestframework      3.14.0
google-api-core          2.11.0
google-api-python-client 2.87.0
google-auth              2.19.0
google-auth-httplib2     0.1.0
google-auth-oauthlib     1.0.0
googleapis-common-protos 1.59.0
httplib2                 0.22.0
idna                     3.4
oauthlib                 3.2.2
pip                      23.0.1
protobuf                 4.23.2
pyasn1                   0.5.0
pyasn1-modules           0.3.0
pyparsing                3.0.9
pytz                     2023.3
requests                 2.31.0
requests-oauthlib        1.3.1
rsa                      4.9
setuptools               66.1.1
six                      1.16.0
sqlparse                 0.4.4
uritemplate              4.1.1
urllib3                  1.26.16
wheel                    0.38.4
